+++ tests/test_physics/test_fock/test_fock.py
@@ -19,22 +19,24 @@ import numpy as np
 from scipy.special import factorial
 from thewalrus.quantum import total_photon_number_distribution
 from mrmustard.lab import (
-    Vacuum,
     Circuit,
+    Vacuum,
     S2gate,
     BSgate,
-    Coherent,
-    SqueezedVacuum,
-    Attenuator,
-    Ggate,
-    Fock,
-    Gaussian,
-    Dgate,
+    Sgate,
     Rgate,
+    Dgate,
+    Ggate,
+    Interferometer,
+    SqueezedVacuum,
     TMSV,
     State,
+    Attenuator,
+    Fock,
+    Coherent,
+    Gaussian,
 )
-from mrmustard.physics.fock import dm_to_ket, ket_to_dm, trace
+from mrmustard.physics.fock import dm_to_ket, ket_to_dm, trace, apply_choi_to_dm
 
 
 # helper strategies
@@ -189,28 +191,52 @@ def test_dm_to_ket_error():
         dm_to_ket(state)
 
 
-def test_fock_trace_mode1():
-    """tests that the Fock state is correctly traced out from mode 1"""
-    state = Vacuum(2) >> Ggate(2)
+def test_fock_trace_mode1_dm():
+    """tests that the Fock state is correctly traced out from mode 1 for mixed states"""
+    state = Vacuum(2) >> Ggate(2) >> Attenuator([0.1, 0.1])
     from_gaussian = state.get_modes(0).dm([3])
-    from_fock = State(dm=state.dm([40])).get_modes(0).dm([3])
+    from_fock = State(dm=state.dm([3, 30])).get_modes(0).dm([3])
     assert np.allclose(from_gaussian, from_fock, atol=1e-5)
 
 
-def test_fock_trace_mode0():
-    """tests that the Fock state is correctly traced out from mode 0"""
-    state = Vacuum(2) >> Ggate(2)
+def test_fock_trace_mode0_dm():
+    """tests that the Fock state is correctly traced out from mode 0 for mixed states"""
+    state = Vacuum(2) >> Ggate(2) >> Attenuator([0.1, 0.1])
     from_gaussian = state.get_modes(1).dm([3])
-    from_fock = State(dm=state.dm([40])).get_modes(1).dm([3])
+    from_fock = State(dm=state.dm([30, 3])).get_modes(1).dm([3])
+    assert np.allclose(from_gaussian, from_fock, atol=1e-5)
+
+
+def test_fock_trace_mode1_ket():
+    """tests that the Fock state is correctly traced out from mode 1 for pure states"""
+    state = Vacuum(2) >> Sgate(r=[0.1, 0.2], phi=[0.3, 0.4])
+    from_gaussian = state.get_modes(0).dm([3])
+    from_fock = State(dm=state.dm([3, 30])).get_modes(0).dm([3])
+    assert np.allclose(from_gaussian, from_fock, atol=1e-5)
+
+
+def test_fock_trace_mode0_ket():
+    """tests that the Fock state is correctly traced out from mode 0 for pure states"""
+    state = Vacuum(2) >> Sgate(r=[0.1, 0.2], phi=[0.3, 0.4])
+    from_gaussian = state.get_modes(1).dm([3])
+    from_fock = State(dm=state.dm([30, 3])).get_modes(1).dm([3])
     assert np.allclose(from_gaussian, from_fock, atol=1e-5)
 
 
 def test_fock_trace_function():
     """tests that the Fock state is correctly traced"""
-    state = Vacuum(2) >> Ggate(2)
-    dm = state.dm([10, 10])
+    state = Vacuum(2) >> Ggate(2) >> Attenuator([0.1, 0.1])
+    dm = state.dm([3, 20])
     dm_traced = trace(dm, keep=[0])
     assert np.allclose(dm_traced, State(dm=dm).get_modes(0).dm(), atol=1e-5)
+
+
+def test_dm_choi():
+    """tests that choi op is correctly applied to a dm"""
+    circ = Ggate(1) >> Attenuator([0.1])
+    dm_out = apply_choi_to_dm(circ.choi([10]), Vacuum(1).dm([10]), [0], [0])
+    dm_expected = (Vacuum(1) >> circ).dm([10])
+    assert np.allclose(dm_out, dm_expected, atol=1e-5)
 
 
 def test_single_mode_choi_application_order():